requirements.txt
pandas
playwright
huggingface_hub
tqdm
pyarrow
python-dotenv
duckdb
